requirements.txt
python-telegram-bot==21.5
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0
